% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmastats.R
\docType{package}
\name{jmastats}
\alias{jmastats}
\title{Access to Japan Meteorological Agency Data}
\description{
Provides the features to download weather data published
by the Japan Meteorological Agency website (\url{http://www.data.jma.go.jp/obd/stats/etrn/index.php}).
}
