% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_unit.R
\name{parse_unit}
\alias{parse_unit}
\title{Parse data variable units}
\usage{
parse_unit(data, rename = TRUE)
}
\arguments{
\item{data}{data}

\item{rename}{\emph{logical}}
}
\value{
a \code{tbl} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
jma_collect("hourly", "0010", 2018, 7, 30) |>
  parse_unit(rename = TRUE)
}
