% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tide_level.R
\name{pivot_tide_level}
\alias{pivot_tide_level}
\title{Convert and split tidal level data}
\usage{
pivot_tide_level(data)
}
\arguments{
\item{data}{tidal level data}
}
\value{
List to store two datasets. hourly observation data and tide.
}
\description{
Convert and split tidal level data
}
\examples{
\dontrun{
read_tide_level("https://www.data.jma.go.jp/gmd/kaiyou/data/db/tide/suisan/txt/2020/TK.txt") \%>\%
  pivot_tide_level()
}
}
\seealso{
\code{\link[=read_tide_level]{read_tide_level()}}
}
