% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kishou_feed.R
\name{read_kishou_feed}
\alias{read_kishou_feed}
\title{Read Kishou Disaster Prevention Information Feed}
\usage{
read_kishou_feed(url)
}
\arguments{
\item{url}{feed url (character)}
}
\value{
data.fram
}
\description{
Read Kishou Disaster Prevention Information Feed
}
\examples{
\dontrun{
read_kishou_feed("http://www.data.jma.go.jp/developer/xml/feed/extra.xml")
}
}
\seealso{
\url{http://xml.kishou.go.jp/index.html}
}
