% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tide_level.R
\name{read_tide_level}
\alias{read_tide_level}
\title{Read and parse tide level text data}
\usage{
read_tide_level(path = NULL, .year, .month, .stn, raw = FALSE)
}
\arguments{
\item{path}{URL or local file path to sea tide level file}

\item{.year}{year}

\item{.month}{month}

\item{.stn}{Two uppercase letters of the alphabet to identify the observation point.}

\item{raw}{If \emph{TRUE}, return raw format data}
}
\description{
Read and parse tide level text data
}
\examples{
\dontrun{
read_tide_level("https://www.data.jma.go.jp/gmd/kaiyou/data/db/tide/suisan/txt/2020/TK.txt")

read_tide_level(.year = 2020, .month = 2, .stn = "TK")
}
}
\seealso{
\url{https://www.data.jma.go.jp/gmd/kaiyou/db/tide/suisan/readme.html}
}
