% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquake_station}
\alias{earthquake_station}
\title{Japan Meteorological Agency's earthquake observe stations}
\format{
A simple feature data frame with 671 rows  7 variables
}
\usage{
earthquake_station
}
\description{
This data corresponds to the January 10, 2024 update.
}
\examples{
head(earthquake_station)
}
\keyword{datasets}
