% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{ggplot2-scales}
\alias{ggplot2-scales}
\alias{scale_color_jma_absolute}
\alias{scale_color_jma_relative}
\alias{scale_fill_jma_absolute}
\alias{scale_fill_jma_relative}
\title{Scales to use for ggplot2}
\usage{
scale_color_jma_absolute(type = "precipitation", ...)

scale_color_jma_relative(type = "amedas", ...)

scale_fill_jma_absolute(type = "precipitation", ...)

scale_fill_jma_relative(type = "amedas", ...)
}
\arguments{
\item{type}{Display item}

\item{...}{Arguments to pass on to \link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn} or
\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn}}
}
\value{
A \code{ScaleContinuous} object that can be added to a \code{ggplot} object

gg scale object
}
\description{
Scales to use for ggplot2
}
\examples{
\donttest{
d <-
jma_collect("daily", block_no = "47646", year = 2017, month = 9, pack = FALSE) |>
parse_unit()

library(ggplot2)
library(units)
ggplot(d, aes(date, precipitation_sum,
          color = drop_units(precipitation_sum),
          fill = drop_units(precipitation_sum))) +
  geom_bar(stat = "identity") +
  scale_color_jma_absolute(type = "precipitation") +
  scale_fill_jma_absolute(type = "precipitation")
}
}
