% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jma_collect.R
\name{jma_collect}
\alias{jma_collect}
\title{Collect JMA Stats Data}
\usage{
jma_collect(item = NULL, block_no, year, month, day)
}
\arguments{
\item{item}{url}

\item{block_no}{block number}

\item{year}{select year}

\item{month}{select month}

\item{day}{select date (default \code{NULL})}
}
\description{
Collect JMA Stats Data
}
\examples{
\dontrun{
jma_collect(item = "annually", "1284", year = 2017, month = 11, day = NULL)
# daily
jma_collect(item = "daily", block_no = "0010", year = 2017, month = 11, day = NULL)
jma_collect(item = "daily", "0422", year = 2017, month = 11, day = NULL)
}
}
