% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jma_collect.R
\name{jma_collect}
\alias{jma_collect}
\title{Collect JMA Stats Data}
\usage{
jma_collect(
  item = NULL,
  block_no,
  year,
  month,
  day,
  cache = TRUE,
  pack = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{item}{url}

\item{block_no}{Block number of the area to be observed. It is assumed that
block_no is input as a string consisting of a 4- or 5-digit number. If a
numeric value is specified, it is processed as a string.}

\item{year}{select year}

\item{month}{select month}

\item{day}{select date (default \code{NULL})}

\item{cache}{use cash and save to cache}

\item{pack}{Whether to packing common variables or not}

\item{quiet}{Whether to output information on variable and
row combinations that were treated as missing values
for some reason. (\code{TRUE}, the default)}
}
\description{
Collect JMA Stats Data
}
\note{
The parameter \code{item} chooses one from these:
\itemize{
\item annually:
\item monthly:
\item 3monthly:
\item 10daily:
\item mb5daily:
\item daily:
\item hourly:
\item rank:
}
}
\examples{
\dontrun{
jma_collect(item = "annually", "1284", year = 2017, month = 11)
# daily
jma_collect(item = "daily", block_no = "0010", year = 2017, month = 11)
jma_collect(item = "daily", "0422", year = 2017, month = 11)
# hourly
jma_collect("hourly", "0010", 2018, 7, 30)
# ranking
jma_collect("rank", block_no = "47646", year = 2020)
}
}
