% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmastats-package.R
\docType{package}
\name{jmastats-package}
\alias{jmastats}
\alias{jmastats-package}
\title{jmastats: Download Weather Data from Japan Meteorological Agency Website}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides features that allow users to download weather data published by the Japan Meteorological Agency (JMA) website (\url{https://www.jma.go.jp/jma/index.html}). The data includes information dating back to 1976 and aligns with the categories available on the website. Additionally, users can process the best track data of typhoons and easily handle earthquake record files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uribo/jmastats}
  \item Report bugs at \url{https://github.com/uribo/jmastats/issues}
}

}
\author{
\strong{Maintainer}: Shinya Uryu \email{suika1127@gmail.com} (\href{https://orcid.org/0000-0002-0493-6186}{ORCID}) [copyright holder]

}
\keyword{internal}
