% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_station.R
\name{nearest_station}
\alias{nearest_station}
\title{Find out neighborhood station}
\usage{
nearest_station(longitude, latitude)
}
\arguments{
\item{longitude}{longitude}

\item{latitude}{latitude}
}
\description{
Find out neighborhood station
}
