% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_unit.R
\name{parse_unit}
\alias{parse_unit}
\title{Parse data variable units}
\usage{
parse_unit(data, rename = TRUE)
}
\arguments{
\item{data}{data}

\item{rename}{*logical*}
}
\description{
Parse data variable units
}
