% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tide_level.R
\name{pivot_tide_level}
\alias{pivot_tide_level}
\title{Convert and split tidal level data}
\usage{
pivot_tide_level(data)
}
\arguments{
\item{data}{tidal level data}
}
\value{
List to store two datasets containing hourly and tide level data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
read_tide_level(system.file("dummy/tide.txt", package = "jmastats")) |>
  pivot_tide_level()
}
\seealso{
\code{\link[=read_tide_level]{read_tide_level()}}
}
