% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eqdb.R
\name{read_eqdb_csv}
\alias{read_eqdb_csv}
\title{Read the csv of the earthquake database}
\usage{
read_eqdb_csv(path, show_metadata = TRUE)
}
\arguments{
\item{path}{local file path to earthquake record file.}

\item{show_metadata}{logical. If \emph{FALSE}, returns only the values
observed at each location.}
}
\value{
a \code{tbl} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
read_eqdb_csv(system.file("dummy/eqdb.csv", package = "jmastats"))
}
\seealso{
\url{https://www.data.jma.go.jp/svd/eqdb/data/shindo/index.html}
}
