% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jma_weather.R
\name{read_jma_weather}
\alias{read_jma_weather}
\title{Read the past weather}
\usage{
read_jma_weather(path)
}
\arguments{
\item{path}{The path to the downloaded file.}
}
\value{
a \code{tbl} object
}
\description{
Read the past weather data files downloaded from JMA.
}
\examples{
read_jma_weather(system.file("dummy/dl_data.csv", package = "jmastats"))
}
\seealso{
\url{https://www.data.jma.go.jp/gmd/risk/obsdl/index.php},
\url{https://www.data.jma.go.jp/gmd/risk/obsdl/top/help3.html}
}
