% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kishou_feed.R
\name{read_kishou_feed}
\alias{read_kishou_feed}
\title{Read Kishou Disaster Prevention Information Feed}
\usage{
read_kishou_feed(frequency, type)
}
\arguments{
\item{frequency}{Release frequency. Select either high frequency ("high") or
long term ("low")}

\item{type}{Feed type. Specify the item to be retrieved as a string.
See details for the items.}
}
\value{
data.frame
}
\description{
Read Kishou Disaster Prevention Information Feed
}
\details{
The following items can be specified in the type argument.
\itemize{
\item regular: It will be announced on time.
\item extra: It will be announced at any time.
\item eqvol: Earthquakes and Volcanoes.
\item other: Other informations.
}
}
\examples{
\dontrun{
read_kishou_feed("http://www.data.jma.go.jp/developer/xml/feed/extra.xml")
}
}
\seealso{
\url{http://xml.kishou.go.jp/index.html}
}
