% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kishou_feed.R
\name{read_kishou_feed}
\alias{read_kishou_feed}
\title{Read Kishou Disaster Prevention Information Feed}
\usage{
read_kishou_feed(frequency, type)
}
\arguments{
\item{frequency}{Release frequency. Select either high frequency ("high") or
long term ("low")}

\item{type}{Feed type. Specify the item to be retrieved as a string.
See details for the items.}
}
\value{
a \code{tbl} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
The following items can be specified in the type argument.
\itemize{
\item regular: It will be announced on time.
\item extra: It will be announced at any time.
\item eqvol: Earthquakes and Volcanoes.
\item other: Other informations.
}
}
\examples{
\donttest{
read_kishou_feed("high", type = "regular")
read_kishou_feed("low", "other")
}
}
\seealso{
\url{https://xml.kishou.go.jp}
}
