% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{JMA's Stations list}
\format{\itemize{
  \item{area}
  \item{station_no}
  \item{station_type}
  \item{station_name}
  \item{address}
  \item{elevation}
  \item{observation_begin}
  \item{note1}
  \item{note1}
  \item{note2}
  \item{katakana}
  \item{prec_no}
  \item{block_no}
  \item{geometry}
}}
\usage{
stations
}
\description{
XXX
}
\keyword{datasets}
