% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Japan Meteorological Agency's Stations list}
\format{
A data frame with 1323 rows 14 variables:
\itemize{
\item{area}
\item{station_no}
\item{station_type}
\item{station_name}
\item{address}
\item{elevation}
\item{observation_begin}
\item{note1}
\item{note1}
\item{note2}
\item{katakana}
\item{prec_no}
\item{block_no}
\item{pref_code}
\item{geometry}
}
}
\usage{
stations
}
\description{
This data corresponds to the April 1, 2024 update.
}
\examples{
head(stations)
dim(stations)
}
\keyword{datasets}
