% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tide_station}
\alias{tide_station}
\title{Tidal observation stations of Japan Meteorological Agency}
\format{
A data frame with 1739 rows 7 variables
}
\usage{
tide_station
}
\description{
XXX
}
\examples{
head(tide_station)
}
\keyword{datasets}
